Fixed a potentially malicious URL in the README
I clicked a link in the README expecting a demo for this app and I was redirected to an oauth screen asking for access to my GitHub my account. On second thoughts it looks like someone just took the instructions a bit too literally, but it's better to be safe than sorry!