use a less problematic word in our intro text
boxen-web was released with some pretty uncool intro text that was
revised somewhat in https://github.com/boxen/boxen-web/pull/7

over at https://github.com/boxen/our-boxen/issues/58 @holman shared his
opinion that while the first paragraph was "cringe-worthy", he doesn't
consider the second paragraph to be distasteful.

i'd like to respectfully claim that both paragraphs were part of the
same problematic reference, and request that we remove all traces of it
in order to set a proper welcoming tone for newcomers to the boxen
community.